Update standards references for 'Basic' authentication
Since moaning about the state of cross-browser compatibility for Basic
authentication on Twitter, I've found the correct current reference for
this -- namely, RFC7617 -- which provides a justification for the
curl/Chrome/any-reasonable-implementer approach of UTF-8 encoding the
user-id and password.

This commit updates the API docs and the comments in the tests to
reflect this.