Only update auth tickets expiry column after a threshold
Before we did an `UPDATE ... RETURNING` query, which updated the auth
tickets `expires` column on every single request. This can lead to
locking problems when one auth ticket is used in a lot of concurrent
requests. Which can lead to the `UPDATE ... RETURNING` query having to
wait until the previous request has been finished and its transaction
committed.

The proposed solution in this commit is to split this up into a `SELECT`
query for checking if the provided ticket is valid, then calculate the
new expiry date, but only set it if the current expiry date is older
with some additional threshold. This threshold is currently set at one
minute.

This also has the effect that in the worst case that users can get
logged out after 6 days, 23 hours, and 59 minutes, instead of 7 days.
This difference however is small enough so that we can ignore it.