Allow submitting an empty search form
We handle key events differently when the search suggestions dropdown is
open. It used to always prevent the default event behaviour, even when
no suggestion from the dropdown was selected, this prevented a user from
submitting an empty form, which can be useful to return to a "search
all" query.

This change only prevents the default key event behaviour if a
suggestion is selected, otherwise it lets the event bubble up and thus
submit the form.

Fixes #3903.