Make SearchBarController tests implementation-agnostic
The existing SearchBarController tests depend heavily on the specific
implementation details of the SearchBarController -- that is, they
assume the use of the 'input' event as opposed to other alternatives
('keyup', 'change', etc.) that would in principle also work.

Coupling tests to implementation in this way can make refactoring very
difficult -- one is forced to change the tests at the same time as the
code. Even more importantly, ensuring correct behaviour in response to
user-triggered events is substantially more complicated than it would at
first appear, because each user event (such as a click) typically
triggers a predefined sequence of related events (in this case,
mousedown, mouseup, focus, click) which need to be simulated as well.

To pick a pathological example to illustrate this point: adding a
'keyup' event handler to the SearchBarController which destroyed the
entire component would clearly break the real controller, but would have
had no effect on this test suite.

In order to address these issues, this commit rewrites the
SearchBarController test suite to use `syn`, a synthetic event library,
which takes care of most of these considerations.

Doing this has exposed at least one bug in the implementation, where the
`handleInputOnInput` function calls `updateActiveDropdownItem` to update
the "active" dropdown item in response to a keypress, and then
immediately calls `maybeOpenOrCloseDropdown`, which (indirectly) calls
`clearActiveDropdownItem`.