Use focus/blur events rather than focusin/focusout
Firefox does not support focusin/focusout, see:

    https://bugzilla.mozilla.org/show_bug.cgi?id=687787

The typical objection to the focus/blur events is that they do not
bubble (although they can be captured with the 'usecapture' parameter to
`addEventListener`). In this case this is not a problem, as we are
listening for the event on the originating target element, so we don't
need to rely on any event bubbling.

Fixes #3894.