Turn search query terms into lozenges as user types.
A phrase is determined to be a valid query term if, when the user
types a space character, the phrase is of the one of the following
forms:
    sometext
    "some text with spaces"
    'some text with spaces'
    <keyword>:sometext
    <keyword>:"some text with spaces"
    <keyword>:'some text with spaces'

After a lozenge is created, pressing the enter button when there are no active
dropdown items highlighted, submits the search form with the query parameter
"q" set to the query terms represented by the lozenges and any unfinished text
within the search box joined with a space character.

A lozenge can be deleted by pressing the "X" button on the right of each lozenge.

When a search page loads any query terms already in the hdden search box are turned
into lozenges.

https://trello.com/c/Iv4mWcIs/19-turn-search-query-terms-into-lozenges-as-user-types