Ensure that moving the mouse above an item reselects it as active
Rewriting the tests in dca2a38 revealed a bug in the updating of active
items in response to mouse events. Specifically: mouseover events don't
come from nowhere, they happen when the mouse cursor crosses the into an
element's bounding rect. Merely moving the cursor above an element does
not trigger a mouseover.

We fix this by updating the active item on mousemove rather than
mouseover.