Don't allow flexbox to shrink annotation cards
Without flex-shrink: 0 the browser will sometimes shrink the annotation
card and make more room for the stats instead (not always - it appears
to depend on the contents of the annotation card and/or the stats). Add
flex-shrink: 0 so that the annotation card always maintains its 500px
width and the stats div gets shrunk instead.