Add OAuth token service
This service will receive an assertion JWT bearer token and a grant_type
and will exchange that for an API token.

Partners will generate JWT tokens for their users and sign them with
their client secret. They then pass that JWT token as a configuration to
the Hypothesis client (that part still needs work), and it will then
exchange the JWT authorization grant for an access token via the `POST
/api/token` endpoint, which will be implemented in the next commit.

It adheres to the behaviour described in RFC7523 [1], except for the
hashing algorithm. RFC7523 requires the use of the RS256 algorithm to
sign the JWT token, the implementation in this commit still uses thi
HS256 algorithm. In the future, authclients would generate a
public/private keypair and share the public part with us.

[1]: https://tools.ietf.org/html/rfc7523