Improve wrapping of URLs and usernames
Improve the wrapping of URLs and usernames in the document buckets on
the search page when they're too wide to fit on one line. Just wrap,
don't attempt to find word boundaries to wrap it, as this can lead to
characters such as ?-_ that may appear in the middle of URLs and
usernames being mistaken for word boundaries, and can create strange
wrapping effects.