Add migration for token.authclient_id column
This way we can track which authclient is responsible for which tokens,
how many active users an authclient has, etc.

It is also useful for the future when everybody will be able to create
authclients, because when an authclient gets deleted all tokens need to
be removed to make sure that they can't be used anymore for making
requests.