Make the `/search` page work without JavaScript.
Fixes #3990.

Get the `/search` page working without JavaScript again.

Previously what happened was this:

1. The server rendered the `/search` page with two `<input>`s in the
search bar `<form>`: an empty, nameless but visible `<input>`, and a
hidden `<input type="hidden" name="q">`. The server rendered any search
text the user had submitted into the hidden input.

2. At controller initialization time, `SearchBarController` would move
the text from the `<input type="hidden" name="q">` into lozenges and/or
into the visible `<input>`.

3. The user would type into the visible input. Some of this text might
be lozengified by `SearchBarController` and moved from the visible input
into lozenges. Some might remain in the visible input. The empty `<input
type="hidden" name="q">` remained empty as the user typed.

4. On form submit `SearchBarController` would read text from the
lozenges and the visible `<input>` and write it into the `<input
type="hidden" name="q">` and then submit the form, so this text
constructed from the lozenges and the visible input would be put into
the hidden input at the last second submitted as the `q` parameter.

The problem is that when JavaScript is not available, the text just
remains in the visible `<input>` and the `<input type="hidden"
name="q">` remains empty, so an empty string is always submitted as the
`q` parameter and search doesn't work. It always submits an empty search
query.

In addition, there is no JavaScript to copy the text from the `<input
type="hidden" name="q">` into the visible `<input>` on page load, so the
user's search query appears to disappear from the search bar on form
submit as well.

What happens now:

1. The server renders the `/search` page with a single, visible `<input
name="q">` containing any search text the user submitted. There is no
hidden input.

2. The user types into the visible input. Some of this text might be
lozengified by `SearchBarController` and moved from the visible input
into lozenges. Some might remain in the visible input.

3. On form submit `SearchBarController` reads text from the lozenges and
the visible `<input>` and writes it into a _new_ `<input type="hidden"
name="q">` which it inserts into the DOM immediately before submitting
the form. `SearchBarController` also moves the `name="q"` HTML attribute
from the visible to the newly-inserted hidden `<input>` at the last
second, before submitting the form.

   (The reason for inserting a hidden input like this rather than just
   setting the value of the visible input is so that users don't see a
   flash as the search text changes for a moment before the page
   reloads.)

When JavaScript is not available, this works exactly like a normal basic
form would do with a single input.

Changes:

* In the server-side HTML template, the `value="{{ q }}"` is moved from
the hidden `<input>` back on to the visible `<input>` where it used to
be.

   This means that, when JavaScript is not available, when the search
   page is reloaded after submitting the search form, the search text
   that the user has typed remains in the visible `<input>` (rendered
   there by the server) instead of disappearing (because the server was
   rendering it into the hidden `<input>` instead).

   As a result of this, `SearchBarController`'s `lozengifyHiddenInput()`
   becomes `lozengifyInput()`, since the search text that it needs to
   lozengify on controller initialization is now in the `<input>` not in
   the hidden `<input>`.

* In the server-side HTML template, the `name="q"` attribute is moved
from the hidden `<input>` back on to the visible `<input>` (where it
used to be). This means that the search bar works without JavaScript
again: the user types text into the visible `<input>` and hits `Enter`
to submit it as the `q` parameter.

* The hidden input is removed from the search bar form (removed from the
server-side HTML template). `SearchBarController` now inserts the hidden
input into the DOM at the last second when it is needed - just before
submitting the form.

* `this._inputHidden` is removed from `SearchBarController`. It just
creates a hidden input when it needs it, doesn't need to keep track of
it in a class-level variable.

* `SearchBarController`'s `updateHiddenInputValue()` becomes
`insertHiddenInput()`, a method that both creates a hidden input and
inserts it into the DOM, _and_ sets the `value` of that hidden `<input>`
to the search query that we want to submit according to the lozenges and
any text in the visible `<input>`.

* A `submitForm()` method is added that inserts the hidden input and
submits the form (called both on `Enter` and on deleting a lozenge)

* `SearchTextParser.getLozengeValues()` now returns an object containing
two properties: the array of lozenge values and the "incomplete input
value" (any un-lozengifiable text, e.g. an unclosed quoted phrase, that
should remain in the `<input>` and not be lozengified).

   This was previously two separate functions `getLozengeValues()` and
   `getIncompleteInputValue()`. This pair of functions is now called in
   one place only - in `lozengifyInput()` - and in that place both the
   lozenge and the remaining text are needed so it makes sense for the
   one function just to return both.

Known issues:

Before this commit there was already a "flash of no lozenges". The
search bar, as rendered by the server, didn't contain any lozenges (or
any search text - it was empty). When `SearchBarController` runs it
inserts the lozenges and search text into the DOM. This results in a
flash where you can see the empty search bar for a moment before the
lozenges and search text are inserted.

By moving the un-lozengified search text, as rendered initially by the
server, from the hidden input into the visible input, this commit would
have made this flash even worse. Instead of seeing an empty `<input>`
for a moment you would see an `<input>` containing un-lozengified text,
until `SearchBarController` lozengifies it.

To avoid this, some CSS involving `env-js-capable` and `env-js-timeout`
has been added to hide the visible input until after lozengification has
happened. This returns us to seeing an empty search bar for a moment
before the lozenges appear.

`visibility: hidden` (rather than, say, `display: none`) is used to hide
the `<input>` so that the input still takes up the same space that it
would have taken, affecting the layout of the other still-visible
elements on the page in the same way. Otherwise (e.g. with `display:
none`), the search bar would shrink in height for a second until the
`<input>` was displayed.