Decouple lozenge tests from each other
Decouple SearchBarController's lozenge tests from each other.

Have each test call a component() helper function in order to
instantiate the SearchBarController instance and its various elements
(e.g. its <input>), rather than having the controller be a global
variable that is set for each test by a beforeEach() function.

This paves the way for parameters to be added to the new component()
function and for different tests to pass different arguments to it, or
for some tests to not use the component() function at all, decoupling
the tests from eachother.

This will enable easy writing of different tests that test the behaviour
of SearchBarController when it is initialized in different contexts (for
example with different values in the <input>).

This same pattern is followed in some other controller tests already,
e.g. character-limit-controller-test.js.