Add "Leave this group" button the group sidebar
Add a "Leave this group" button to the group sidebar on the
/groups/<pubid>/search page.

The button is _not_ shown if the authenticated user is the group's
creator, because group creators can't leave groups.

A `<button formmethod="POST">` is used for the _Leave this group_ "link" so
that clicking on it does a `POST` request instead of a `GET`. This is
semantically correct because because the request actually changes
something in the model (removes the user from the group's membership).

Use the HTML `form` attribute on this `<button>` so that the
`<button>` is considered to be part of the search bar form
(the `<form id="search-bar">`) which is further up on the page, even
though the `<button>` element is not inside this `<form>` element in the
hierarchical HTML structure. This ensures that any query parameters in
the search bar form's `<input>`, including ones that have been typed
into the input but not submitted to the server yet (so they aren't in
the URL yet), are submitted and not lost when the leave group button is
clicked.

Use of the `form` attribute means that the leave group button will not
work in any version of Internet Explorer, which doesn't support the
`form` attribute. This is to be fixed in a follow-up commit.

Add an invisible (as in CSS `display:none;`) `<input type="submit">` to
the search bar `<form>` at the top of the page. The form did not have
any submit element (it has to be submitted by hitting the Enter key).
When you hit Enter the browser will try to find the nearest submit
element in the form and, if it can find one, will submit the form as if
you had clicked that button. Previously the browser didn't find any such
submit element for the search bar form. Adding the new leave group
button meant that hitting Enter in the search box acted as if you had
clicked the leave group button! So add an invisible submit input to
prevent this.