Fix layout of search page when no search results
Now that we have a sidebar to the right of the search results on the
search page (at least when it's a /groups/<pubid>/search URL) when there
are no search results (e.g. a group that has no annotations yet) the
sidebar would be drawn on the left hand side of the window instead. Make
sure that the empty search results container still takes up space, so
that the sidebar appears in the right postion.