Add "More info" / "Back" button
On the /groups/<pubid>/search page, hide the sidebar and show only the
search results on small screens.

Add a navigation component showing the group name and a "More info"
button to the top of the page on small screens.

The nav shows a "More info" button that, when clicked, loads a version
of the page in which the nav shows a "Back" button instead, and the
sidebar is shown instead of the search results.