Add username and group pubid from url as lozenges to the nav bar.
User and Group profile pages are not adding the facet to the search box on load.
This is because a query 'q' from which lozenges are created and which contains a
single group or user term is redirected to the specific group or user search page
with that term removed.

https://github.com/hypothesis/product-backlog/issues/16