Add group invite component to group sidebar
Create a new "panel" (a reusable collection of Python logic, Jinja2 HTML
template code and CSS, implemented using the "panels" mechanism of the
pyramid_layout package) for a group invite component and use it in the
group sidebar on the /groups/<pubid>/search page.

The Jinja2 template code for this panel is duplicated between the new
panel (in group_invite.html.jinja2) and the old group page
(share.html.jinja2). I didn't want to change the old group page to use
the new panel because, unfortunately, it requires slightly different
HTML. For example share.html.jinja2 uses <p
class="group-form-footer__heading"> elements for its titles (!), whereas
I wanted to use an <h3> element for the title in the new component. It's
also arguably not a good idea for legacy and new pages to share code.