Make clicking usernames work
On the /groups/<pubid>/search page make clicking on one of the usernanes
in the sidebar add that username as a facet to the current search query
(and reload the page, updating the search results).

If the username is _already_ in the search query, then clicking on the
username in the sidebar _removes_ it from the query.