Redirect group overview page to group search page
For users with the search_page feature enabled, redirect requests for
the legacy group overview page to the new group profile page which is a
variant of the search page.

This results in the navbar's dropdown group links and links to the
group's page from the client land on the new group profile page.