Don't lose unsubmitted search query terms
In some cases unsubmitted search query terms were being lost on page
load. For example, go to the /groups/<pubid>/search page, type "foo"
into the search box but don't submit it yet (so "foo" is in the search
form, but is not yet in the URL), then click on a tag in the sidebar to
add a facet for that tag into the search query.

The page reloads and the tag facet has been added into the query, but
"foo" (and any unsubmitted query terms, whether they had been
"lozengified" by JavaScript or not) has been lost from the query.

The problem is that code is accessing the search query parameter `q`
from request.params, which is the params in the URL. It should be
accessing it from request.POST, which is the params from the HTML form.

In the tests, by default DummyRequest.params and DummyRequest.POST are
**the same object**. So if a test does, for example,
DummyRequest.POST['q'] = 'foo' this will set both DummyRequest.POST['q']
and DummyRequest.params['q'], if it then tests that "foo" is preserved
in the URL that a view function redirects to, this will test will pass
even though the code is incorrectly accessing DummyRequest.params
instead of DummyRequest.POST.

The fix is to set DummyRequest.POST and DummyRequest.params to two
different objects in the tests, so that they can catch this bug.