Fix error loading unorm
Fix an error when `normalize()` tries to require the 'unorm' bundle on
certain routes. This function is called on every page that displays the
navbar but the bundle was only included on /search routes.

This commit changes unorm to be loaded in the same way as other
polyfills and for the moment, included on every page.

In future we can be smarter and load the polyfill only in browsers that
actually need it, since the majority of our users have browsers which
support String.prototype.normalize() natively.