Fix sidebar buttons in IE
Various buttons in the search page sidebar and nav (tags, usernames,
leave group, more info, and back) rely on the `form` attribute which IE
doesn't support.

Make sure that the entire search-result-nav and search-result-sidebar
are wrapped in a <form> element. For IE the buttons inside these
elements will belong to the new <form>s. For other browsers that read
the `form` attribute the buttons will continue to belong to the
search-bar form.

Some of these buttons were already wrapped in <form>s for IE, others
weren't. Wrapping the entire nav and sidebar reduces the number of forms
needed and makes sure that all buttons are caught.

This makes clicking on the buttons work in IE.

Known issues: unlike in other browsers, in IE any unsubmitted query
words in the search bar will be lost when clicking on any of these
buttons. Already-submitted words will be preserved.

Fixes #4077.