Autofocus search input when user types a letter, number or backspace
Copy the behavior seen on Google.com where the search field is
autofocused when the user types an alphanumeric character or presses
Backspace, if no other input on the page has focus.

This allows the user to start typing a search query as soon as the page loads
without having to focus the input field first.