Use consistent DOM structure for client/server rendered lozenges
The client-side implementation of lozenges used a different
DOM-structure than server-rendered lozenges. This caused several
problems:

 - The client JS code for lozenges could not be used with
   server-rendered lozenges.

 - The client DOM structure failed at accessibility by using a <div>
   rather than a <button> for the close button - hence it did not
   support keyboard focus, identify as a button for screen readers etc.

 - LozengeController did not follow the patterns and conventions of
   other controllers, which all work by taking an existing
   server-rendered element as input and enhancing it with behavior.

Fix this by rendering a <template> for client-generated lozenges in the
markup from the server and cloning that as the basis for new lozenges in
client code.

Sharing the structure should also make it easier to render lozenges on
the server in future to avoid the FOUC on initial page load.