Instrument elasticsearch query requests in Search client
We have two functions that send a request to ElasticSearch, this commit
adds an optional instrumentation code path to these two functions. When
the Search client gets initialized with a stats client, then we'll time
how long it takes the request to finish, and also add counters for
successful, timed-out, and errored search queries.