Add plain and in-context links to the document in the bucket sidebar
Make the document URI in a bucket a link and add a link to see the
bucket's annotations in context. This is currently a direct link to the
first annotation in the bucket since the direct linking infrastructure
does not support direct links to documents without specifying a
particular annotation.

Fixes #4058