Choose login redirect URL based on feature flags for logged-in user
Previously the URL to redirect to after login was determined at the
start of request processing before authentication completed and
consequently it used the feature flags for _logged out_ users.

This makes the request redirect based on the flags for the logged in
user by deferring retrieval of the login URL until after login and
clearing the feature flag cache just before the login is processed.

Fixes #4070