Match facet types case insensitive in search bar
We show a list of tag or group suggestions when a user types in "tag:"
or "group:". But this doesn't work when the user types in "Tag:" or
"GROUP:". This commit makes sure that we match the facet type case
insensitive by lowercasing the string before we match that it starts
with "tag:" or "group:".

Fixes #4080.