Fix removing current tag facet from tag sidebar
When the current search query includes "tag:foo tag:bar" and the user is
clicking on the tag "foo" in the sidebar (which toggles the tag), we end
up with a search query that looks like: "tag:[u'bar']".

This is because we're building up a list of tag facets, removing the one
the user wants to remove, and then set the list of remaining tags on the
parameters (which is a MultiDict). What we need to do is loop through
the list of remaining tags and add them one by one to the MultiDict
parameters.

Fixes #4076.