Allow reindexing into a new index (one-pass)
This commit implements a simplistic one-pass strategy for reindexing
into a *new* index (rather than on top of the existing index) if we
detect that the open index is behind an alias.

This will need to be refined before it can be used in production, as
currently updates made to annotations while the reindex is running will
be lost.