Move logic for generating in-context links for buckets into DocumentBucket
Move the logic for generating in-context links to document buckets out
of the template and into the DocumentBucket class. This makes it easier
to add comments and tests.

The dict containing the presentable form of a bucket annotation has been
moved from DocumentBucket.annotations to
DocumentBucket.presented_annotations because a) it was confusing that
DocumentBucket.annotations had different types at different points in
the request and b) it prevented DocumentBucket from being able to rely
on the type of that property.