Fix removing current user facet from group members sidebar
When the current search query includes "user:fred user:mary" and the user is
clicking on the username "fred" in the sidebar (which toggles the user facet),
we end up with a search query that looks like: "user:[u'mary']".

This is because we're building up a list of user facets, removing the one
the user wants to remove, and then set the list of remaining user facets on the
parameters (which is a MultiDict). What we need to do is loop through
the list of remaining user facets and add them one by one to the MultiDict
parameters.

Fixes #4101.