Don't index annotation "extra" field data
As convenient as it may be to allow people to attach arbitrary data to
annotations and subsequently search by those fields, it's fundamentally
problematic due to the nature of the inverted index that Elasticsearch
builds for our documents.

Allowing end users to insert arbitrary keys into indexed documents means
that they can quite straightforwardly perform a denial-of-service attack
against Hypothesis by adding thousands of new keys to annotations.

In addition, the already-inflated size of our index mapping is one
suspect in the mysterious slow query problems we've been having
recently.