Check requested against available version for static assets
If a request for a static asset specifies a version, check that it
matches the available version as specified in the JSON manifest. If it
does not, return a 404.

This fixes the problem where a client can request a particular version
of an asset but actually receive a newer version - which may be a
problem for certain clients.

Keeping old versions of assets available is not handled by the service
but must be handled by a CDN or other caching proxy in front of it.