Make a single Elasticsearch client per application
Previously, we made an Elasticsearch instance per-request, which is
appealing from an isolation point of view but unfortunate from a
resource usage and performance standpoint because the connection pool is
associated with the client instance.

(There is no distinction between an "engine" and a "session" as in
SQLAlchemy, which makes some sense because the client is in theory
stateless.)

This commit changes the code so that a single client is created on
startup, and reused in every request. This should make a noticeable
improvement on query latency, because we won't be creating and tearing
down the TCP connections and TLS sessions with Elasticsearch for every
single request.