Turn get_aliased_index/update_aliased_index into helpers
This commit adds get_aliased_index and update_aliased_index as helper
functions that receive a client instance. These are currently
implemented as methods on `memex.search.client.Client`, but this
probably isn't wise when the `Client` instance lives for the life of the
application -- every method on that class is an opportunity to leak
state between requests.

Subsequent commits will update those parts of the code that use this
functionality and remove the old implementation.