Create aliased index with initdb or autoconfig
In the development environment the application will automatically create
a search index if it doesn't find one. In production the same behaviour
can be triggered by a call to

    hypothesis initdb

This commit updates the code so that it creates a randomly named index
(the name is based off the supplied index name) such as

    hypothesis-abcd1234

and creates an alias for the supplied index name pointing to the
concrete index. This makes it much easier to reindex into a new index
without having to redeploy or restart the application.

This commit also removes quite a bit of code that is no longer needed
now that "documents" aren't stored in the search index, and adds tests
for the config functionality.