Fix FOUC in lozenges on page load
Use an inline SVG rather than an async-loaded image for the lozenge
delete icon to avoid a layout flicker and flash of <img> alt text on
page load.

The padding tweak is needed to ensure correct vertical centering of the
icon tested in Firefox and Chrome.

Move the label for the action onto the button itself.