Remove trailing /search from group search page
Make the group search page use the /groups/{pubid}/{slug} URL instead of
/groups/{pubid}/search. This means it uses the same URL as the "Log in
to join {group}" and the "Join {group}" pages (whose views are
implemented in h.views.groups:

* Not logged in: /groups/{pubid}/{slug} shows the "Log in to join {group}" page
* Logged in but not a member of the group: shows the "Join {group}" page
* Already a member of the group: shows the group's activity search page
* No group with the {pubid} exists: 404s

The /groups/{pubid}/search route (route name: "activity.group_search")
no longer exists.

The other behavioral change in this commit is that the /search page will
no longer redirect to the groups page if you search for a group that
doesn't exist. This is the same as it does for users.