Redirect to right group slug if missing or wrong
When the search_page feature flag is off, or when the user isn't a
member of the group or isn't logged in, the 'group_read' views in
h.views.groups redirect requests with no or the wrong slug to the correct
slug.

For example /groups/{pubid}, /groups/{pubid}/ and /groups/{pubid}/foobar all
redirect to /groups/{pubid}/{slug}.

Make h.views.activity.GroupSearchController, which is used for the
'group_read' route when the user is logged in, is a member of the group,
and has the search_page feature flag on, do the same redirecting.