Show classic groups page if feature flag off
If the "search_page" feature flag is off then show the classic group
page, not the new group activity search page, at
`/groups/<pubid>/<slug>`.

When the group activity search page was moved from `/groups/<pubid>/search` to
`/groups/<pubid>/<slug>` it started to be shown even if the
"search_page" feature flag is off. This fixes that.