Move responsibility for JSON parsing into received_message
This moves the responsibility for parsing incoming message JSON into the
`WebSocket` instance's `received_message` method. Now that we are
closing the connection in response to invalid JSON this seems like a
more reasonable place for it as it ensures we close connections sending
bad data as soon as possible.