Add a reply method to websocket.Message
This commit adds a reply method to websocket.Message which sends a
response to the client, prefilling appropriate `reply_to` and `ok`
fields.

For backwards compatibility with our existing client (which never sends
an `id` field with its messages) we simply drop the response on the
floor if the incoming message is lacking an appropriate id.