Report errors for invalid messages from the client
This commit adds error handling to the websocket API. Messages from the
client which are invalid will be replied to, but ONLY if those messages
contain a message ID.

This change should be entirely backwards compatible, as the error
message replies won't be sent to our existing client, which doesn't set
message IDs.

The commit adds tests for these error paths, and moves the existing
tests for these functions into classes as in a082415.