Use traversal for /users/{username} pages (#4136)
Use Pyramid's traversal mechanism to find the user model object for the
/users/{username} route, as we do for the /groups/{pubid}/{slug} route.

This commit introduces a couple of behavioral changes:

1. Searching for user:does_not_exist on the /search page will stay on
the /search page (the URL will become /search?q=user:does_not_exist).
The server no longer redirects the browser to /users/does_not_exist.

2. /users/does_not_exist now 404s for users that don't exist in the db

3. Searching for user:username_that_does_exist on the /search page still
redirects you to /users/username_that_does_exist as before. The view
function for this page (UserSearchController.search()) no longer needs
to look up the User model object in the database itself, because Pyramid
passes this object to the view as the context object. The view also no
longer needs to handle the case where the user doesn't exist, because
Pyramid's view lookup will now 404 the page and not call the view at all
in this case.

UserFactory is added in h.models.user for Pyramid traversal to use. This
is the same as GroupFactory in h.models.group (I'm not sure that a
traversal resource belongs in a models file but I'm putting it in the
same place as GroupFactory).

The add_route() for the 'activity.user_search' route is changed to use
the `factory` and `traverse` arguments, the same as the `group_read`
route.