Use traversal for group search views
Use Pyramid's traveral mechanism, rather than URL dispatch, to select
the view for the "activity.group_search" route, the same as is already
done for "activity.user_search" and for "group_read" and other group
routes.

This means that `/groups/<pubid_that_does_not_exist>/search` will not
404.

This means that the views in `GroupSearchController` no longer need to
look up the `Group` object in the database themselves (it is passed to
them as `self.group`), nor handle the case where the `Group` doesn't
exist (in that case Pyramid will 404 without ever calling the view).