Don't redirect to group pages that don't exist
Now that `/groups/<pubid_that_does_not_exist>/search` is a 404 instead
of an empty search results page, the `/search` page should not redirect
to it if the user submits a search for
`group:pubid_that_does_not_exist`.

This is consistent with the behaviour of the
`/users/<username_that_does_not_exist>` page and searching for
`user:username_that_does_not_exist`.