Factor out GroupUserSearchController
Replace it with private helper methods instead. This is a refactoring
with no functional change. It should make it easier to make future
changes to these view classes by allowing view defaults to be specified
more easily.

`GroupUserSearchController` was a class containing view methods that each
had two `@view_config`s on: one for the `"activity.group_search"` route
and one for `"activity.user_search"`. This avoided duplicating the Python
code in the view methods, which is shared by the group and user search
routes.

The problem with this is that it ends up duplicating view config
declarations - `@view_defaults` can't be used effectively on
`GroupUserSearchController` because it can only contain defaults that
apply to both the group search and the user search routes for every view
method in the class. Additionally the view defaults needed for the group
search and user search routes are already on `@view_defaults`
declarations on `GroupSearchController` and `UserSearchController`
respectively.

Instead move all of `GroupUserSearchController`'s methods to private
helper functions: `_more_info(request)` etc. Then for each of the
private helper functions add a corresponding method to _both_
`GroupSearchController` _and_ `UserSearchController`:
`GroupSearchController.more_info(self)` etc. These duplicated methods
are just one-liners that call the helper methods (avoiding duplication
of Python code) and that have `@view_config`s on them. These
`@view_config`s can rely on the `@view_defaults` of
`GroupSearchController` and `UserSearchController`, avoiding duplication
of view config declarations.

We can now add more view defaults to `UserSearchController` or
`GroupSearchController` separately and have them apply to all the view
methods they contain.